requirements.txt
prometheus-client
git+https://github.com/ollytheninja/PyFlick.git

aiohttp~=3.6.2
python-dotenv~=0.19.0